fix event list
